refactor: main화면 home화면으로 변경
